Changed SampleBuffer to use all 16 bit values so that struct packing doesn't make my life hell

+++ src/logger.h
@@ -49,7 +49,7 @@ typedef struct RingBuffer
 typedef struct SampleBuffer
 {
     uint16_t adc[ADC_CHANNELS];
-    int8_t accel[ACCEL_CHANNELS];
+    uint16_t accel[ACCEL_CHANNELS];
 } SampleBuffer;
 
 void logger_init(void);
